Fix remaining GenFunction s32→intptr_t parameter mismatches in function definitions

+++ src/core2/code_CBBF0.c
@@ -12,7 +12,7 @@ typedef struct{
     s32 unk10;
 }CommonParticleType;
 
-void commonParticleType_set(enum common_particle_e id, GenFunction_0 init_method, s32 update_method, GenFunction_0 free_method, s32 arg4, s32 arg5);
+void commonParticleType_set(enum common_particle_e id, GenFunction_0 init_method, intptr_t update_method, GenFunction_0 free_method, intptr_t arg4, s32 arg5);
 
 /* .bss */
 CommonParticleType D_80386260[0x12];
@@ -26,7 +26,7 @@ void commonParticleType_init(void){
     }
 }
 
-void commonParticleType_set(enum common_particle_e id, GenFunction_0 init_method, s32 update_method, GenFunction_0 free_method, s32 arg4, s32 arg5){
+void commonParticleType_set(enum common_particle_e id, GenFunction_0 init_method, intptr_t update_method, GenFunction_0 free_method, intptr_t arg4, s32 arg5){
     D_80386260[id].init_method = init_method;
     D_80386260[id].update_method = update_method;
     D_80386260[id].free_method = free_method;
